add overlay to view modifier
