+++ Sources/Playbook/Design Elements/Position/PositionCatalog.swift
@@ -11,12 +11,11 @@ import SwiftUI
 
 public struct PositionCatalog: View {
   public var body: some View {
-    VStack(alignment: .leading) {
+    VStack(alignment: .leading, spacing: 0) {
       PBAvatar(image: Image("andrew", bundle: .module), size: .large)
-        .overlay {
+        .position(top: 35, left: 0, bottom: 0, right: 0) {
           PBBadge(text: "On Roadtrip", rounded: true, variant: .neutral)
             .background(Color.card)
-            .positioning(top: 35, left: 0, bottom: 0, right: 0)
         }
     }
     .navigationTitle("Position")
